Move namada::types::error into namada::sdk::error
this type is used quite extensively in the SDK and the shared crate in
general, any change to it should be noted in the changelog under SDK
changes as it is developer facing
