+++ apps/src/lib/client/rpc.rs
@@ -34,6 +34,8 @@ use namada::ledger::pos::{CommissionPair, PosParams, Slash};
 use namada::ledger::queries::RPC;
 use namada::ledger::storage::ConversionState;
 use namada::proof_of_stake::types::{ValidatorState, WeightedValidator};
+use namada::sdk::error;
+use namada::sdk::error::{is_pinned_error, Error, PinnedBalanceError};
 use namada::sdk::masp::{
     Conversions, MaspAmount, MaspChange, ShieldedContext, ShieldedUtils,
 };
@@ -44,13 +46,12 @@ use namada::sdk::rpc::{
 use namada::sdk::wallet::{AddressVpType, Wallet};
 use namada::types::address::{masp, Address};
 use namada::types::control_flow::ProceedOrElse;
-use namada::types::error::{is_pinned_error, Error, PinnedBalanceError};
 use namada::types::hash::Hash;
 use namada::types::key::*;
 use namada::types::masp::{BalanceOwner, ExtendedViewingKey, PaymentAddress};
 use namada::types::storage::{BlockHeight, BlockResults, Epoch, Key, KeySeg};
 use namada::types::token::{Change, MaspDenom};
-use namada::types::{error, storage, token};
+use namada::types::{storage, token};
 use tokio::time::Instant;
 
 use crate::cli::{self, args};
