+++ shared/src/sdk/mod.rs
@@ -7,5 +7,6 @@ pub mod signing;
 #[allow(clippy::result_large_err)]
 pub mod tx;
 
+pub mod error;
 pub mod queries;
 pub mod wallet;
