+++ shared/src/sdk/signing.rs
@@ -28,6 +28,7 @@ use crate::ibc::applications::transfer::msgs::transfer::MsgTransfer;
 use crate::ibc_proto::google::protobuf::Any;
 use crate::ledger::parameters::storage as parameter_storage;
 use crate::proto::{MaspBuilder, Section, Tx};
+use crate::sdk::error::{EncodingError, Error, TxError};
 use crate::sdk::masp::{
     make_asset_type, ShieldedContext, ShieldedTransfer, ShieldedUtils,
 };
@@ -43,7 +44,6 @@ use crate::sdk::tx::{
 pub use crate::sdk::wallet::store::AddressVpType;
 use crate::sdk::wallet::{Wallet, WalletUtils};
 use crate::sdk::{args, rpc};
-use crate::types::error::{EncodingError, Error, TxError};
 use crate::types::key::*;
 use crate::types::masp::{ExtendedViewingKey, PaymentAddress};
 use crate::types::storage::Epoch;
