+++ shared/src/types/mod.rs
@@ -1,7 +1,6 @@
 //! Types definitions.
 
 pub mod control_flow;
-pub mod error;
 pub mod ibc;
 pub mod key;
